Move common function to util class
